Refactor url functions into new file
