+++ pkg/scraper/xpath.go
@@ -2,26 +2,15 @@ package scraper
 
 import (
 	"bytes"
-	"context"
 	"errors"
-	"fmt"
-	"io"
-	"net/http"
-	"net/http/cookiejar"
 	"net/url"
 	"regexp"
 	"strings"
 	"time"
 
 	"github.com/antchfx/htmlquery"
-	"github.com/chromedp/cdproto/dom"
-	"github.com/chromedp/cdproto/network"
-	"github.com/chromedp/chromedp"
-	jsoniter "github.com/json-iterator/go"
 
 	"golang.org/x/net/html"
-	"golang.org/x/net/html/charset"
-	"golang.org/x/net/publicsuffix"
 
 	"github.com/stashapp/stash/pkg/logger"
 	"github.com/stashapp/stash/pkg/models"
@@ -119,61 +108,9 @@ func (s *xpathScraper) scrapeSceneByFragment(scene models.SceneUpdateInput) (*mo
 }
 
 func (s *xpathScraper) loadURL(url string) (*html.Node, error) {
-	var r io.Reader
-
-	driverOptions := s.config.DriverOptions
-	if driverOptions != nil && driverOptions.UseCDP {
-		// get the page using chrome dp
-		resp, err := urlFromCDP(url, *driverOptions)
-		if err != nil {
-			return nil, err
-		}
-		r = strings.NewReader(resp)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		// get the page using http.Client
-		options := cookiejar.Options{
-			PublicSuffixList: publicsuffix.List,
-		}
-		jar, er := cookiejar.New(&options)
-		if er != nil {
-			return nil, er
-		}
-
-		client := &http.Client{
-			Timeout: scrapeGetTimeout,
-			// defaultCheckRedirect code with max changed from 10 to 20
-			CheckRedirect: func(req *http.Request, via []*http.Request) error {
-				if len(via) >= 20 {
-					return errors.New("stopped after 20 redirects")
-				}
-				return nil
-			},
-			Jar: jar,
-		}
-
-		req, err := http.NewRequest("GET", url, nil)
-		if err != nil {
-			return nil, err
-		}
-
-		userAgent := s.globalConfig.UserAgent
-		if userAgent != "" {
-			req.Header.Set("User-Agent", userAgent)
-		}
-
-		resp, err := client.Do(req)
-		if err != nil {
-			return nil, err
-		}
-		defer resp.Body.Close()
-
-		r, err = charset.NewReader(resp.Body, resp.Header.Get("Content-Type"))
-		if err != nil {
-			return nil, err
-		}
+	r, err := loadURL(url, s.config, s.globalConfig)
+	if err != nil {
+		return nil, err
 	}
 
 	ret, err := html.Parse(r)
@@ -248,88 +185,4 @@ func (q *xpathQuery) subScrape(value string) mappedQuery {
 	}
 
 	return q.scraper.getXPathQuery(doc)
-}
-
-// func urlFromCDP uses chrome cdp and DOM to load and process the url
-// if remote is set as true in the scraperConfig  it will try to use localhost:9222
-// else it will look for google-chrome in path
-func urlFromCDP(url string, driverOptions scraperDriverOptions) (string, error) {
-	if !driverOptions.UseCDP {
-		return "", fmt.Errorf("Url shouldn't be feetched through CDP")
-	}
-
-	remote := false
-	sleep := 2
-
-	if driverOptions.Remote {
-		remote = true
-	}
-
-	if driverOptions.Sleep != 0 {
-		sleep = driverOptions.Sleep
-	}
-
-	sleepDuration := time.Duration(sleep) * time.Second
-	act := context.Background()
-
-	if remote {
-		var cancelAct context.CancelFunc
-		remote, errCDP := getRemoteCDP()
-		if errCDP != nil {
-			return "", errCDP
-		}
-		act, cancelAct = chromedp.NewRemoteAllocator(context.Background(), remote)
-		defer cancelAct()
-	}
-
-	ctx, cancel := chromedp.NewContext(act)
-	defer cancel()
-
-	var res string
-	err := chromedp.Run(ctx,
-		network.Enable(),
-		chromedp.Navigate(url),
-		chromedp.Sleep(sleepDuration),
-		chromedp.ActionFunc(func(ctx context.Context) error {
-			node, err := dom.GetDocument().Do(ctx)
-			if err != nil {
-				return err
-			}
-			res, err = dom.GetOuterHTML().WithNodeID(node.NodeID).Do(ctx)
-			return err
-		}),
-	)
-	if err != nil {
-		return "", err
-	}
-
-	return res, nil
-}
-
-// func getRemoteCDP returns the complete remote address that is required to access the cdp instance
-func getRemoteCDP() (string, error) {
-	resp, err := http.Get("http://localhost:9222/json/version")
-	if err != nil {
-		return "", err
-	}
-
-	var result map[string]interface{}
-	var json = jsoniter.ConfigCompatibleWithStandardLibrary
-	if err := json.NewDecoder(resp.Body).Decode(&result); err != nil {
-		return "", err
-	}
-	remote := result["webSocketDebuggerUrl"].(string)
-	logger.Debugf("Remote cdp instance found %s", remote)
-	return remote, err
-}
-
-func cdpNetwork(enable bool) chromedp.Action {
-	return chromedp.ActionFunc(func(ctx context.Context) error {
-		if enable {
-			network.Enable().Do(ctx)
-		} else {
-			network.Disable().Do(ctx)
-		}
-		return nil
-	})
 }
